backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.9
botocore==1.35.9
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
cssselect2==0.7.0
daphne==4.1.2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-storages==1.14.4
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ffmpeg-python==0.2.0
filelock==3.16.1
future==1.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jmespath==1.0.1
lxml==5.3.0
msgpack==1.1.0
oscrypto==1.3.0
pbr==6.1.0
pillow==10.3.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pypdf==5.0.0
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
razorpay==1.4.2
redis==5.1.0
reportlab==4.2.2
requests==2.32.3
s3transfer==0.10.2
secure-smtplib==0.1.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
stevedore==5.3.0
svglib==1.5.1
tinycss2==1.3.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
virtualenv==20.26.6
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zope.interface==7.0.3
